Changes half.h to use other halffloat library and added basic NifNodeTree, plus others
(Uncompilable until I fix large amount of if's problem in NifNodeTree)
(removed qhull folder for now)

+++ NvTriStrip/NvTriStripObjects.h
@@ -227,7 +227,7 @@ namespace NvTriStrip
 		void RemoveSmallStrips(NvStripInfoVec& allStrips, NvStripInfoVec& allBigStrips, NvFaceInfoVec& faceList);
 
 		bool FindTraversal(NvFaceInfoVec &faceInfos, NvEdgeInfoVec &edgeInfos, NvStripInfo *strip, NvStripStartInfo &startInfo);
-		int  CountRemainingTris(std::list<NvStripInfo*>::iterator iter, std::list<NvStripInfo*>::iterator  end);
+		size_t CountRemainingTris(std::list<NvStripInfo*>::iterator iter, std::list<NvStripInfo*>::iterator  end);
 
 		void CommitStrips(NvStripInfoVec &allStrips, const NvStripInfoVec &strips);
